refactor: 리팩토링 #8

+++ Back-End/src/main/java/com/conseller/conseller/core/gifticon/api/dto/response/ExpiringGifticonResponse.java
@@ -8,5 +8,5 @@ public final class ExpiringGifticonResponse {
     private final long userIdx;
     private final int expiryDay;
     private final String gifticonName;
-    private final int gifticonCnt;
+    private final long gifticonCnt;
 }
